Finish Readme on dev environment, clean up endpoints, remove check_task endpoint, tests passing
