+++ dev/requirements.txt
@@ -1,0 +1,71 @@
+amqp==2.3.2
+appnope==0.1.0
+atomicwrites==1.2.1
+attrs==18.2.0
+backcall==0.1.0
+billiard==3.5.0.4
+bleach==2.1.4
+celery==4.2.1
+certifi==2018.10.15
+chardet==3.0.4
+Click==7.0
+coverage==4.5.1
+decorator==4.3.0
+defusedxml==0.5.0
+entrypoints==0.2.3
+falcon==1.4.1
+Flask==1.0.2
+freezegun==0.3.10
+gunicorn==19.9.0
+html5lib==1.0.1
+idna==2.7
+ipykernel==5.0.0
+ipython==7.0.1
+ipython-genutils==0.2.0
+ItsDangerous==1.0.0
+jedi==0.13.0
+Jinja2==2.10
+jsonschema==2.6.0
+kombu==4.2.1
+MarkupSafe==1.0
+mistune==0.8.3
+mock==2.0.0
+more-itertools==4.3.0
+nbconvert==5.4.0
+nbformat==4.4.0
+numpy==1.15.2
+pandas==0.23.4
+pandocfilters==1.4.2
+parso==0.3.1
+pbr==5.1.0
+pexpect==4.6.0
+pickleshare==0.7.5
+Pillow==5.3.0
+pluggy==0.8.0
+prometheus-client==0.3.1
+prompt-toolkit==2.0.5
+ptyprocess==0.6.0
+py==1.7.0
+Pygments==2.2.0
+pytest==3.9.2
+python-dateutil==2.7.3
+python-mimeparse==1.6.0
+pytz==2018.5
+pyzmq==17.1.2
+redis==2.10.6
+requests==2.20.0
+scikit-learn==0.20.0
+scipy==1.1.0
+Send2Trash==1.5.0
+simplegeneric==0.8.1
+six==1.11.0
+sklearn==0.0
+terminado==0.8.1
+testpath==0.4.2
+tornado==5.1.1
+traitlets==4.3.2
+urllib3==1.24
+vine==1.1.4
+wcwidth==0.1.7
+webencodings==0.5.1
+Werkzeug==0.14.1
